Rename some files Change class to function comps apply use state Remove unused files
