+++ src/components/Board/board.js
@@ -1,0 +1,38 @@
+import React from 'react';
+import Square from '../Square';
+import './board.css';
+
+const Board = (props) => {
+  const { squares, onClick } = props;
+
+  const renderSquare = (i) => {
+    return (
+      <Square
+        value={squares[i]}
+        onClick={() => onClick(i)}/>
+    )
+  }
+
+  return (
+    <div>
+      <div className="board-row">
+        {renderSquare(0)}
+        {renderSquare(1)}
+        {renderSquare(2)}
+      </div>
+      <div className="board-row">
+        {renderSquare(3)}
+        {renderSquare(4)}
+        {renderSquare(5)}
+      </div>
+      <div className="board-row">
+        {renderSquare(6)}
+        {renderSquare(7)}
+        {renderSquare(8)}
+      </div>
+    </div>
+  );
+
+}
+
+export default Board;
