+++ src/components/Square/index.js
@@ -1,11 +1,3 @@
-import React from 'react';
-import './square.css';
+import Square from './square'
 
-const Square = (props) => {
-    return (
-        <button className="square" onClick={props.onClick}>
-            {props.value}
-        </button>
-    )
-}
-export default Square;
+export default Square
